Converted to using express router
